// desktop/src/open-api/model/customField.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomFieldOption } from './customFieldOption';
import { CustomFieldType } from './customFieldType';


export interface CustomField { 
    id: number;
    createdAt: string;
    createdBy?: number;
    /**
     * Created by entity\'s name
     */
    createdByString?: string;
    updatedAt?: string;
    /**
     * Custom Field name
     */
    name: string;
    type: CustomFieldType;
    /**
     * Custom Field description
     */
    description?: string;
    options?: Array<CustomFieldOption>;
}
export namespace CustomField {
}